Time: 0 ms (100.00%), Space: 5.9 MB (31.21%) - LeetHub

+++ 7-reverse-integer/7-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int n) {
-        long long int ans = 0;
+        int ans = 0;
         while(n!=0)
         {
             int digit = n% 10;
